+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Upgrade to SWT v4.18
 - Upgrade to JRuby v9.2.14.0
 - Apply all WidgetProxy property converters upon normal setting of properties too (not just in DSL) (like `some_widget.background = color_symbol`)
+- Update Hello, Menu Bar! sample to show accelerators on menu items
+- Have the `swt` keyword (SWTProxy) support accepting a string character (to build an accelerator style)
+- Make accelerator property accept symbols and character directly (without swt)
+- Fix enablement on `menu` (as opposed to menu_item, where it works)
 
 ### 4.17.10.8